umqtt: Improve compatibility with CPython
On CPython, the file object returned by "socket.makefile()" does not
offer a ".write()" method obtaining a "length" argument. Apparently,
the MicroPython documentation does not mention that parameter as well,
so it is probably an undocumented optimization.

To make things work on CPython, this makes ".write()" obtain a single
argument while the data is sliced by using "memoryview".