(u)urequests: Include port number into "Host" header.
When responding with a relative redirect, most HTTP servers will use
the value of the "Host" header to expand this to an absolute URL
before populating the "Location" header. When the port number is
missing, things will go haywire when the HTTP client tries to perform
the redirect.

The HTTP specification says that the port number must be included
when the port is not the default port per scheme (80 for http://
and 443 for https://). However, it probably does not do any harm
when always including it.